WIP - About Page
content.json - added "titleRotate" to the paragraphs in the about section. Will be used when outputting the content as in About.jsx (paragraph-title-rotate${paragraph.titleRotate})

About.jsx - added <h3> and title. Modified the names of the names of classes that rotate the text to be dynamic - paragraph-title-rotate${paragraph.titleRotate} . These correspond to the renamed classes in globals.css.

globals.css - renamed classes paragraph-title-why to paragraph-title-rotate60, paragraph-title-how to paragraph-title-rotate300, paragraph-title-who to paragraph-title-rotate120, and paragraph-title-what to paragraph-title-rotate200