Update analytics.js
Just the ID and we're good to go.
Note: Because this is a slightly different setup for the Layout component, I'm not sure if the pages will be tracked individually but let's just see what happens. If this only logs the main page, we can fix that quickly

Let's do the npm audit in a separate PR. Feel free to go for it if you like. :)
components/utils/analytics.js
import ReactGA from 'react-ga';

 export const initGA = () => {
  ReactGA.initialize('UA-113130296-3');

 @atlisg atlisg 3 days ago  Member
This is the FOOD-E ID, the correct ID is UA-113130296-2